refactored code, created an abstract game class
